+++ Cargo.toml
@@ -3,6 +3,7 @@ panic = "unwind"
 
 [workspace]
 members = [
+	"primitives",
 	"node",
 	"runtime/rococo",
 	"standalone/node",
