+++ primitives/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+authors = ["Webb Technologies Inc."]
+edition = "2021"
+license = "Unlicense"
+name = "tangle-primitives"
+version = "1.0.0"
+
+[dependencies]
+smallvec = "1.6.1"
+frame-support = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.29" }
+sp-runtime = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.29" }
+sp-core = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.29" }
+sp-io = { git = 'https://github.com/paritytech/substrate.git', default-features = false, branch = "polkadot-v0.9.29" }
+
+[features]
+default = ["std"]
+std = [
+    "frame-support/std",
+    "sp-runtime/std",
+    "sp-core/std",
+    "sp-io/std",
+]
