+++ runtime/rococo/Cargo.toml
@@ -7,6 +7,7 @@ version = "3.0.0"
 homepage = "https://webb.tools"
 repository = "https://github.com/webb-tools/tangle"
 edition = "2021"
+
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
 
@@ -127,7 +128,7 @@ pallet-author-inherent = { git = "https://github.com/webb-tools/nimbus", branch 
 
 # Local dependencies
 pallet-parachain-staking = { path = '../../pallets/parachain-staking', default-features = false }
-# tangle-primitives = { path = '../../primitives', default-features = false }
+tangle-primitives = { path = '../../primitives', default-features = false }
 
 [features]
 default = ["std"]
@@ -224,6 +225,7 @@ std = [
 	# local
 	"pallet-ecdsa-claims/std",
 	"pallet-parachain-staking/std",
+	"tangle-primitives/std"
 ]
 runtime-benchmarks = [
 	"hex-literal",
